+++ chapter_5/matrix_multiplication_shared_mem.cu
@@ -13,9 +13,9 @@
  */
 __global__
 void MatrixMultiplicationKernel(
-    int* matrix_M,
-    int* matrix_N,
-    int* matrix_Out,
+    float* matrix_M,
+    float* matrix_N,
+    float* matrix_Out,
     int Width
 ) {
     // Initialize the shared memories.
@@ -33,7 +33,7 @@ void MatrixMultiplicationKernel(
     int Col = block_x * blockDim.x + thread_x;
 
     // Loop over the tiles required to compute matrix_Out elements.
-    int out_value = 0;
+    float out_value = 0;
     for (int phase = 0; phase < ceil(Width*1.0/TILE_WIDTH); ++phase) {
         // Collaboratively load M and N tiles into shared memory.
         int in_matrix_M_index = Row*Width + phase*TILE_WIDTH + thread_x;
@@ -63,16 +63,16 @@ void MatrixMultiplicationKernel(
 
 
 void runMatrixMultiplication(
-    int* matrix_M_h,
-    int* matrix_N_h,
-    int* matrix_Out_h,
+    float* matrix_M_h,
+    float* matrix_N_h,
+    float* matrix_Out_h,
     int Width
 ) {
     // Get size in bytes.
-    size_t size = Width * Width * sizeof(int);
+    size_t size = Width * Width * sizeof(float);
 
     // Load and copy matrix M and N to device memory.
-    int * matrix_M_d, * matrix_N_d, * matrix_Out_d;
+    float * matrix_M_d, * matrix_N_d, * matrix_Out_d;
     cudaMalloc((void***)&matrix_M_d, size);
     cudaMemcpy(matrix_M_d, matrix_M_h, size, cudaMemcpyHostToDevice);
 
@@ -101,9 +101,9 @@ int main() {
     int Width = 9;
 
     // Define identical matrices M and N where each element = 1 .. Width * Width.
-    int * matrix_M = (int *) malloc(Width * Width * sizeof(int));
-    int * matrix_N = (int *) malloc(Width * Width * sizeof(int));
-    int * matrix_Out = (int *) malloc(Width * Width * sizeof(int));
+    float * matrix_M = (float *) malloc(Width * Width * sizeof(float));
+    float * matrix_N = (float *) malloc(Width * Width * sizeof(float));
+    float * matrix_Out = (float *) malloc(Width * Width * sizeof(float));
 
     for (int i = 1; i <= Width * Width; ++i) {
         matrix_M[i - 1] = i;
@@ -114,7 +114,7 @@ int main() {
 
     for (int i = 0; i < Width; ++i) {
         for (int j = 0; j < Width; ++j) {
-            printf("%d ", matrix_Out[i * Width + j]);
+            printf("%.2f ", matrix_Out[i * Width + j]);
         }
         printf("\n");
     }
